Move around some code
